requirements.txt
attrs==23.2.0
Babel==2.10.3
bcrypt==3.2.2
blinker==1.7.0
certifi==2023.11.17
click==8.1.6
cryptography==41.0.7
Jinja2==3.1.2
jsonschema==4.10.3
MarkupSafe==2.1.5
pydantic==1.10.14
Pygments==2.17.2
PyJWT==2.7.0
PyNaCl==1.5.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
typing_extensions==4.10.0
urllib3==2.0.7
matplotlib
mysql-connector